poc(loadbalancer): add new load balancing strategy
This is a proof-of-concept change to add the "power of two random
choices" load balancing strategy to traefik.

Currently the only load balancing strategy available within traefik is
(weighted) round robin. While this is a good strategy in general, it can
be problematic if one or more of the backends is getting overloaded.

There are many load balancing strategies, but the
"power-of-two-random-choices" algorithm has some good properties that
make it a good general use algorithm.

Specifically some of the benefits include:
 - reducing the total number of requests to slower backends
 - constant time backend picking (in the general case)
 - reduced "herd behaviour" compared to e.g. "least connections" load
   balancing.

The algorithm works by taking two backends at random, and choosing the
backend that has the fewest in-flight requests. In order to do this, we
have to track the number of in-flight requests each backend is currently
processing.

The aim of this change is to demonstrate that this new load balancing
strategy can be added with minimal changes, and reusing a lot of the
existing load balancing code by factoring out the explicit strategy into
an interface.

In order to do this, the wrr package was removed, and the existing
LoadBalancer was moved to the parent directory: the loadbalancer
package.

There are many strategies that can be used for load balancing, many of
which require "extrinsic" information, such as the CPU load on the
backends. This change is not meant to open the door for adding such
strategies, but attempts to add an effective load balancing strategy
with low cost to the codebase.

This change is a PoC and is meant for demonstration purposes. There
would need to be more tests added, updates to documentation, and perhaps
some investigation into performance / optimisations. I am willing to do
this work, but didn't want to spend too much time on this if the change
is not going to be accepted.